Move ZIS services to separate sources.
Resolves #13
