+++ h/zis/service.h
@@ -4,27 +4,32 @@
   This program and the accompanying materials are
   made available under the terms of the Eclipse Public License v2.0 which accompanies
   this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
-  
+
   SPDX-License-Identifier: EPL-2.0
-  
+
   Copyright Contributors to the Zowe Project.
 */
 
-#ifndef ZIS_VERSION_H_
-#define ZIS_VERSION_H_
+#ifndef ZIS_SERVICE_H_
+#define ZIS_SERVICE_H_
 
-#define ZIS_VERSION "1.0.4"
+#define RC_ZIS_SRVC_OK                  0
+#define RC_ZIS_SRVC_STATUS_NULL         2
+#define RC_ZIS_SRVC_SERVICE_FAILED      4
+#define RC_ZIS_SRVC_CMS_FAILED          8
 
-#endif /* ZIS_VERSION_H_ */
+#define ZIS_MAX_GEN_SRVC_RC             32
+
+#endif /* ZIS_SERVICE_H_ */
 
 
 /*
   This program and the accompanying materials are
   made available under the terms of the Eclipse Public License v2.0 which accompanies
   this distribution, and is available at https://www.eclipse.org/legal/epl-v20.html
-  
+
   SPDX-License-Identifier: EPL-2.0
-  
+
   Copyright Contributors to the Zowe Project.
 */
